requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
openai==0.27.8
chromadb==0.3.21
sentence-transformers==2.2.2
numpy==1.24.3
lxml==4.9.3
validators==0.22.0
